# Improve scheduler's performance by eliminating sorting of nodes by their score