@stewart-yu 2018-06-23
“replaces `the sort` with a `linear scan`.”, LGTM